wip(lib/api): refactor and rewrite API endpoint logic
Addresses #122; starts refactoring and rewriting API endpoint logic into
reusable "component" functions.

Starts by rewriting the create match API endpoint in
`lib/api/create-match.ts`.
