+++ lib/api/deprecated/helpers/truncation.ts
