+++ lib/api/deprecated/helpers/verify.ts
