+++ lib/api/helpers/error.ts
@@ -1,11 +1,14 @@
-import { NextApiResponse } from 'next';
-
-export default function error(
-  res: NextApiResponse,
-  msg: string,
-  code = 400,
-  err?: Error
-): void {
-  console.error(`[ERROR] Sending client ${code} with msg (${msg})...`, err);
-  res.status(code).send({ msg, ...(err || {}) });
+/**
+ * Custom `Error` class that contains an HTTP status code property.
+ * @typedef {Object} APIError
+ * @property code - The HTTP status code for this error (e.g. 401 or 500).
+ * @todo Add i18n to human-readable error codes.
+ * @todo Replace `code` with `status` and add a string `code` property that
+ * allows our front-end to know exactly what error has occurred (e.g. a `code`
+ * indicating that the user's JWT is invalid could be `auth/no-jwt`).
+ */
+export default class APIError extends Error {
+  public constructor(message: string, public readonly code: number = 400) {
+    super(message);
+  }
 }
