+++ lib/api/time-in-availability.ts
@@ -1,0 +1,8 @@
+export default function timeInAvailability(
+  req: Req,
+  res: Res,
+  people: UserWithRoles[],
+  time: Timeslot
+): boolean {
+  return false;
+}
